requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.0
huggingface-hub==0.29.3
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
langchain==0.3.22
langchain-community==0.3.20
langchain-core==0.3.49
langchain-text-splitters==0.3.7
tavily-python==0.5.2
pymongo==4.11.3
openai==1.69.0
python-dotenv==1.1.0
python-multipart==0.0.9
matplotlib==3.8.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.32.3
pillow==10.1.0
snowflake-connector-python==3.14.0
snowflake-snowpark-python==1.29.1
PubChemPy==1.0.4
rdkit==2023.9.5
httpx==0.28.1
pandas==2.2.3